Replace Docker edge agent with LXD agent for turbine simulation, WSL Ubuntu 22.04 for ABI compat

+++ cpp/agent/src/signal_generator.cpp
@@ -41,7 +41,7 @@ std::vector<double> generateImpulseTrain(double faultFreq, double shaftSpeed, in
 }
 
 std::vector<double> generateNoise(int sampleRate, double noiseLevel, double durationSeconds) {
-    std::vector<double> signal(sampleRate * durationSeconds);
+    std::vector<double> signal(static_cast<size_t>(sampleRate * durationSeconds));
     std::mt19937 gen(std::random_device{}());
     std::normal_distribution<double> dist(0.0, noiseLevel);
     for (int i = 0; i < signal.size(); i++) {
